# docs/openapi/external-analysis-v1.yaml
openapi: 3.1.0
info:
  title: Developer Analytics External Analysis API
  version: 1.0.0
  description: |
    Compact, user-scoped analytics API intended for external GPT/LLM clients.

    Read endpoints require a Developer Analytics external-client bearer token and
    the versioned Accept media type:
    `application/vnd.developer-analytics.analysis.v1+json`.

    Returned AI assessments use the same bearer token and require the
    `AI_ASSESSMENTS_WRITE` endpoint scope.

    Endpoint scopes and privacy scopes are independent. The server enforces both.
servers:
  - url: https://developer-analytics.example.com
    description: Replace with the deployed Developer Analytics base URL
security:
  - externalClientBearer: []
...
paths:
...
  /api/me/ai-assessments:
    post:
      tags: [AI assessment write-back]
      operationId: returnAiAssessment
      summary: Return an external AI assessment
      description: |
        Requires `AI_ASSESSMENTS_WRITE`.
        `sourceClient`, timestamp and data scope are derived from the authenticated
        token and cannot be supplied by the caller.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAiAssessment'
      responses:
        '200':
          description: Stored assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiAssessment'
        '400':
          description: Invalid assessment payload
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: |
            Missing write scope, or `containsPrivateData=true` with a
            `PUBLIC_ONLY` privacy scope.
components:
  securitySchemes:
    externalClientBearer:
      type: http
      scheme: bearer
      bearerFormat: DeveloperAnalyticsExternalToken
      description: |
        User-created `da_ext_...` token. It is separate from browser sessions
        and GitHub credentials.
...
  responses:
    Unauthorized:
      description: Missing, invalid or revoked external-client token.
...
  schemas:
...
    PrivacyScope:
      type: string
      enum:
        - PUBLIC_ONLY
        - PUBLIC_PLUS_PRIVATE_AGGREGATES
        - FULL_AUTHORISED_ANALYSIS
...
    CreateAiAssessment:
      type: object
      required:
        - analysisType
        - content
        - containsPrivateData
      properties:
        analysisType:
          type: string
          minLength: 1
          maxLength: 120
          example: developer-profile
        content:
          type: object
          additionalProperties: true
          description: Structured AI-generated assessment content.
        containsPrivateData:
          type: boolean
    AiAssessment:
      type: object
      required:
        - id
        - analysisType
        - sourceClient
        - timestamp
        - dataScope
        - content
        - containsPrivateData
      properties:
        id:
          type: string
          format: uuid
        analysisType: { type: string }
        sourceClient: { type: string }
        timestamp:
          type: string
          format: date-time
        dataScope:
          $ref: '#/components/schemas/PrivacyScope'
        content:
          type: object
          additionalProperties: true
        containsPrivateData:
          type: boolean